requirements.txt
amqp==5.0.9
asgiref==3.5.0
attrs==21.4.0
Babel==2.9.1
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coverage==6.3.1
Deprecated==1.2.13
Django==3.2.12
django-environ==0.8.1
django-extra-views==0.13.0
django-filter==21.1
django-haystack==3.1.1
django-oscar==3.1
django-phonenumber-field==5.2.0
django-rest-framework==0.1.0
django-smart-load-tag==0.3.2
django-tables2==2.3.4
django-treebeard==4.4
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
factory-boy==3.2.1
Faker==12.1.0
idna==3.3
iniconfig==1.1.1
kombu==5.2.3
packaging==21.3
phonenumbers==8.12.42
Pillow==9.0.1
pluggy==1.0.0
prompt-toolkit==3.0.28
purl==1.6
py==1.11.0
pyparsing==3.0.7
pysolr==3.9.0
pytest==7.0.1
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2021.3
redis==4.1.4
requests==2.27.1
six==1.16.0
sorl-thumbnail==12.7.0
sqlparse==0.4.2
tomli==2.0.1
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
